btoproject

+++ src/boundary/BTOProjectMain.java
@@ -3,6 +3,11 @@ package boundary;
 import controller.BTOProjectController;
 import entity.project.BTOProject;
 import entity.user.Manager;
+import enums.FlatType;
+import enums.UserRole;
+import java.time.LocalDate;
+import java.time.format.DateTimeParseException;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Scanner;
 
@@ -10,11 +15,20 @@ public class BTOProjectMain {
     PrintProjects printer = new PrintProjects();
     BTOProjectController projectController = new BTOProjectController();
     
+    public static void main(String[] args) {
+        Manager manager = new Manager("123", "password", UserRole.MANAGER);
+        BTOProjectMain btoProjectMain = new BTOProjectMain();
+        btoProjectMain.displayMenu(manager);
+    }
+
     public void displayMenu(Manager manager) {
         try (Scanner sc = new Scanner(System.in)) {
-            int choice;
+            int choice = 0;
             do { 
                 System.out.print("""
+                    ---------------------------
+                      Project Management Menu
+                    ---------------------------
                     1. Create project
                     2. Edit project
                     3. Delete project
@@ -22,17 +36,29 @@ public class BTOProjectMain {
                     5. View my projects
                     6. Exit
                     """);
-                choice = sc.nextInt();
+                boolean validInput = false;
+                while (!validInput) {
+                    try {
+                        System.out.print("Option: ");
+                        choice = sc.nextInt();
+                        sc.nextLine();
+                        validInput = true;
+                    } catch (Exception e) {
+                        System.out.println("Invalid input. Please enter a number.");
+                        sc.nextLine(); 
+                    }
+                }
+
                 switch (choice) {
                     case 1 -> {
-                        createProject(manager);
+                        createProject(manager, sc);
                     }
                     case 2 -> {
-                        editProject(manager);
+                        editProject(manager, sc);
                         break;
                     }
                     case 3 -> {
-                        deleteProject(manager);
+                        deleteProject(manager, sc);
                         break;
                     }
                     case 4 -> {
@@ -55,61 +81,150 @@ public class BTOProjectMain {
         }
     }
 
-    private void createProject(Manager manager) {
-        try (Scanner sc = new Scanner(System.in)) {
-            System.out.print("Project name: ");
-            String projectName = sc.nextLine();
-            if (projectController.projectExist(projectName)) {
-                System.out.println("Project already exists.");
-                return;
+    private void createProject(Manager manager, Scanner sc) {
+        // Project name
+        System.out.print("Project name: ");
+        String projectName = sc.nextLine();
+        if (projectController.projectExist(projectName)) {
+            System.out.println("Project already exists.");
+            return;
+        }
+        // Project neighbourhood
+        System.out.print("Project neighbourhood: ");
+        String neighbourhood = sc.nextLine();
+        // Types of Flat + Number of units
+        Map<FlatType, Integer> unitCounts = new HashMap<>();
+        for (FlatType flatType : FlatType.values()) {
+            int units = 0;
+            while (units <= 0) {
+                try {
+                    System.out.printf("Number of units for %d-Room flats: ", flatType.getNumRooms());
+                    units = sc.nextInt();
+                    if (units <= 0) {
+                        System.out.println("Number of units must be greater than 0.");
+                    }
+                } catch (Exception e) {
+                    System.out.println("Invalid input. Please enter a number.");
+                    sc.nextLine(); 
+                } 
             }
-
-            System.out.print("Project neighbourhood: ");
-            String neighbourhood = sc.nextLine();
-            System.out.print("""
-                    Flat Type:
-                    2 for Two-Room
-                    3 for Three-Room
-                    Enter number: 
-                    """);
-            int flatType = sc.nextInt();
-
-            BTOProject createdProject = projectController.createProject(manager, projectName, neighbourhood);
-            if (createdProject == null) {
-                System.out.println("Could not create project.");
-            } else {
-                System.out.printf("Project %s successfully created!", projectName);
+            sc.nextLine();
+            unitCounts.put(flatType, units);
+        }
+        // Application opening date
+        LocalDate oDate = null;
+        while (oDate == null) {
+            System.out.print("Application opening date (YYYY-MM-DD): ");
+            String oDateInput = sc.nextLine();
+            try {
+                oDate = LocalDate.parse(oDateInput);
+            } catch (DateTimeParseException e) {
+                System.out.println("Invalid date.");
             }
+        }
+        // Application closing date
+        LocalDate cDate = null;
+        while (cDate == null) {
+            System.out.print("Application closing date (YYYY-MM-DD): ");
+            String cDateInput = sc.nextLine();
+            try {
+                cDate = LocalDate.parse(cDateInput);
+            } catch (DateTimeParseException e) {
+                System.out.println("Invalid date.");
+            }
+        }
+        // Available HDB Office Slots
+        int slots = 0;
+        while (slots <= 0) {
+            try {
+                System.out.printf("Available HDB Office Slots: ");
+                slots = sc.nextInt();
+                if (slots <= 0) {
+                    System.out.println("Number of slots must be greater than 0.");
+                }
+            } catch (Exception e) {
+                System.out.println("Invalid input. Please enter a number.");
+                sc.nextLine(); 
+            } 
+        }
+
+        BTOProject createdProject = projectController.createProject(manager, projectName, neighbourhood, unitCounts, oDate, cDate, slots);
+        if (createdProject == null) {
+            System.out.println("Could not create project.");
+        } else {
+            System.out.printf("Project %s successfully created!\n", projectName);
         }
     }
 
-    private void editProject(Manager manager) {
-        try (Scanner sc = new Scanner(System.in)) {
-            System.out.print("Enter name of project to edit: ");
-            String projectName = sc.nextLine();
-            Map<Integer, String[]> options = Map.of(
-                1, new String[]{"Project Name", "NAME"},
-                2, new String[]{"Project neighbourhood", "NEIGHBOURHOOD"},
-                3, new String[]{"Project flat type", "FLAT_TYPE"},
-                4, new String[]{"Number of units", "NUM_UNITS"},
-                5, new String[]{"Application opening date", "OPENING_DATE"},
-                6, new String[]{"Application opening date", "CLOSING_DATE"}
-            );
-            System.out.println("What would you like to edit?");
-            options.forEach((key, value) -> System.out.println(key + ". " + value[0]));
+    private void editProject(Manager manager, Scanner sc) {
+        if (manager.getManagedProjects().isEmpty()) {
+            System.out.println("No projects to edit.");
+            return;
+        }
+        System.out.print("Enter name of project to edit: ");
+        String projectName = sc.nextLine();
 
+        int choice = 0;
+        while (choice != 7) {
+            System.out.println("""
+            What would you like to edit?
+            1. Project Name
+            2. Project neighbourhood
+            3. Number of units
+            4. Application opening date
+            5. Application closing date
+            6. Toggle Visibility
+            7. Exit edit project menu
+                """);
             System.out.print("(Enter number) Attribute to edit: ");
-            int choice = sc.nextInt();
-            sc.nextLine();
-
-            String attribute = options.get(choice)[1];
-            if (attribute == null) {
-                System.out.println("Invalid choice");
+            choice = sc.nextInt();
+            boolean edited = false;
+            switch (choice) {
+                case 1 -> {
+                    System.out.print("New project name: ");
+                    String newProjectName = sc.nextLine();
+                    edited = projectController.editProjectName(manager, projectName, newProjectName);
+                }
+                case 2 -> {
+                    System.out.print("New neighbourhood: ");
+                    String newNeighbourhood = sc.nextLine();
+                    edited = projectController.editNeighbourhood(manager, projectName, newNeighbourhood);
+                }
+                case 3 -> {
+                    System.out.print("Flat type to edit (2 or 3): ");
+                    int flatTypeInput = sc.nextInt();
+                    FlatType flatType = FlatType.getFlatType(flatTypeInput);
+                    if (flatType == null) {
+                        System.out.println("Invalid flat type.");
+                    }
+                    System.out.print("New number of units: ");
+                    int newNumUnits = sc.nextInt();
+                    edited = projectController.editNumUnits(manager, projectName, flatType, newNumUnits);
+                }
+                case 4 -> {
+                    System.out.print("New application opening date (YYYY-MM-DD): ");
+                    String oDateInput = sc.nextLine();
+                    LocalDate oDate = LocalDate.parse(oDateInput);
+                    edited = projectController.editOpeningDate(manager, projectName, oDate);
+                }
+                case 5 -> {
+                    System.out.print("New application closing date (YYYY-MM-DD): ");
+                    String cDateInput = sc.nextLine();
+                    LocalDate cDate = LocalDate.parse(cDateInput);
+                    edited = projectController.editClosingDate(manager, projectName, cDate);
+                }
+                case 6 -> {
+                    System.out.print("Toggle visibility (true/false): ");
+                    boolean visible = sc.nextBoolean();
+                    edited = projectController.editVisibility(manager, projectName, visible);
+                }
+                case 7 -> {
+                    System.out.println("Exiting edit project menu.");
+                }
+                default -> {
+                    System.out.println("Invalid option.");
+                }
             }
-            System.out.print("Enter new value: ");
-            String value = sc.nextLine();
-
-            boolean edited = projectController.editProject(manager, projectName, attribute, value);
             if (edited) {
                 System.out.println("Project updated!");
                 // print update project details ?
@@ -119,16 +234,19 @@ public class BTOProjectMain {
         }
     }
 
-    private void deleteProject(Manager manager) {
-        try (Scanner sc = new Scanner(System.in)) {
-            System.out.print("Delete project: ");
-            String projectName = sc.nextLine();
-            boolean deleted = projectController.deleteProject(manager, projectName);
-            if (deleted) {
-                System.out.printf("Project %s successfully deleted!\n", projectName);
-            } else {
-                System.out.println("Project not found");
-            }
+    private void deleteProject(Manager manager, Scanner sc) {
+        if (manager.getManagedProjects().isEmpty()) {
+            System.out.println("No projects to delete.");
+            return;
         }
+        System.out.print("Delete project: ");
+        String projectName = sc.nextLine();
+        boolean deleted = projectController.deleteProject(manager, projectName);
+        if (deleted) {
+            System.out.printf("Project %s successfully deleted!\n", projectName);
+        } else {
+            System.out.println("Project not found");
+        }
+        
     }
 }
